[FIX] html_editor: add background color indicator for icons
Before this commit: the background color isn't properly indicated in the
toolbar.

After this commit: we create a provider for icon's background color, and
extract the common code in the color_plugin as an utils function.
